Refactors AddMenu to use local state and accept a title
