+++ components/AddMenu/icons/index.js
@@ -1,0 +1,15 @@
+const add = require('./add.png')
+const text = require('./text.png')
+const link = require('./link.png')
+const image = require('./image.png')
+const camera = require('./camera.png')
+const cancel = require('./cancel.png')
+
+export default {
+  add,
+  text,
+  link,
+  image,
+  camera,
+  cancel,
+}
