+++ state/actions.js
@@ -1,8 +1,6 @@
 export const SET_CURRENT_ROUTE = 'SET_CURRENT_ROUTE'
 export const SET_HEADER_TITLE_VISIBILITY = 'SET_HEADER_TITLE_VISIBILITY'
 export const SET_HEADER_MENU_VISIBILITY = 'SET_HEADER_MENU_VISIBILITY'
-export const TOGGLE_ADD_MENU = 'TOGGLE_ADD_MENU'
-export const CLOSE_ADD_MENU = 'CLOSE_ADD_MENU'
 export const SEND_ALERT = 'SEND_ALERT'
 export const DISMISS_ALERT = 'DISMISS_ALERT'
 export const DISMISS_ALL_ALERTS = 'DISMISS_ALL_ALERTS'
